refactor(components.ng-ui-router-line-progress): renamed package name
replace package name from `@ovh-ux/ng-ui-router-line-progress` to `@ovh-ux/ng-ui-router-line-progress`
and also renamed module name from `ngUirouterLineProgress` to `ngUiRouterLineProgress`

BREAKING CHANGE: package and module are now respectively named as
`@ovh-ux/ng-ui-router-line-progress` and `ngUiRouterLineProgress`
