+++ packages/components/ng-ui-router-line-progress/examples/index.html
