+++ packages/components/ng-ui-router-line-progress/src/ng-ui-router-line-progress.config.js
