+++ packages/manager/apps/public-cloud/package.json
@@ -36,7 +36,7 @@
     "@ovh-ux/ng-translate-async-loader": "^2.0.0",
     "@ovh-ux/ng-ui-router-breadcrumb": "^0.1.0",
     "@ovh-ux/ng-ui-router-layout": "^1.2.1",
-    "@ovh-ux/ng-uirouter-line-progress": "^0.1.0",
+    "@ovh-ux/ng-ui-router-line-progress": "^0.1.0",
     "@uirouter/angularjs": "^1.0.20",
     "angular": "^1.7.5",
     "angular-animate": "^1.7.5",
